requirements.txt
robotframework-seleniumlibrary==4.1.0
robotframework==3.1.2
robotframework-faker==4.3.0
robotframework-pabot==1.2.1
robotframework-datadriver==0.3.6
selenium==3.141.0
pydevd==1.9.0
webdriver-manager==2.3.0
psutil==5.7.0
pywin32==227
ImageHash==4.0
PyAutoIt==0.6.0
xlrd==1.2.0
pandas==1.0.3
numpy==1.18.4